Moved fast type 1 check to own class and printing results correctly
